Move _ChainThread to ChainController
